chore: move compiled files to dist directory to make vitest convertion easier (#32406)
