+++ cli/tsconfig.esm.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "rootDir": "./lib",
+    "outDir": "./dist",
+    "target": "ES2022",
+    "module": "ES2022",
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "resolveJsonModule": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "noImplicitAny": false,
+  },
+  "include": [
+    "lib/**/*.mts",
+  ]
+}
