+++ cli/tsconfig.json
@@ -1,5 +1,7 @@
 {
   "compilerOptions": {
+    "rootDir": "./lib",
+    "outDir": "./dist",
     "target": "es2016",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
     "module": "commonjs",                                /* Specify what module code is generated. */
     "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
@@ -7,16 +9,8 @@
     "strict": true,                                      /* Enable all strict type-checking options. */
     "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
     "noImplicitAny": false,
-    "types": [
-      "mocha"
-    ]
   },
   "include": [
-    "index.ts",
-    "lib/**/*.ts"
-  ],
-  "exclude": [
-    "types",
-    "scripts"
+    "lib/**/*.ts",
   ]
 }
