+++ scripts/binary/trigger-publish-binary-pipeline.js
@@ -10,6 +10,7 @@ const { getNextVersionForBinary } = require('../get-next-version')
   const { nextVersion } = await getNextVersionForBinary()
 
   const body = JSON.stringify({
+    branch: 'chore/account_for_dist_move_cli',
     parameters: {
       temp_dir: os.tmpdir(),
       sha: process.env.CIRCLE_SHA1,
